Create a version of JSON throwing unchecked exceptions
Hundreds of lines of codes in 122 files simply deals with catching
JSONException and transforming it into a RuntimeException.

This commit introduce JSON arrays and objects whose method directly
throws JSONException_, a non-checked exception, instead of
JSONException.

Finally, it introduces JSONTokener_, which is similar to JSONTokener,
but creates Array_ and Object_ instead of Array and Object.

For each method `m` which may throws an exception in JSONObject and
JSONArray there is a method `m_`, similar to `m`, which throws
JSONException_ instead of JSONException. Furthermore, each time
getArray or getObject is used, it returns a an Array_ or an Object_
respectively.

When the code is non trivial, it is based on the original
implementatino of JSONObject/JSONArray/JSONTokener.

The original methods are kept identical. So when the code catch the
exception and do something more interesting with it than throwing it
as a RuntimeExpression, the original method can still be used and the
exception will be catched.

This is particularly useful when there is a real risk that the JSON is broken