New model can be edited without full sync
dae/anki@0ffb3080043a9547fd041ca54c8f67a03c96e592,
dae/anki@9863b6f9ff2d75763bcc055395a35a8cb1f43a19 and
dae/anki@b78480fe5282fd54ed56b98252e197cfc2aaff10

This commit solves a problem that I had many time in the past. When I
create a new model, I usually want to edit it. Clone of existing
models present no interest by themselves. And as soon as I edit it, I
need to do a full sync.

As far as I understand ankiweb (which is sadly closed source), the
full sync is required because ankiweb needs to know that the model
associated to note type on the server did change. But since the model
is new, it has no note type associated to on the server, so there is
no need to do a full sync immediatly. Since the model is new, it also
means there is no risk of the inconsistency with a change made in
another computer/smartphone.

Thus, when a field/template is added, I check that the model is not
new by checking both whether it's id is not null, and also that it's
usn is not -1. (I set usn early in the model's life)