Factorize Sched
Avoiding an empty constructor and using the same variable (through
getter/setter) allows to have an easier automated testing of Sched;
since it ensures that each variables are actually set when the object
exists.

Furthermore, it ensures that a SchedV2 object accesing a private
variable access the correct one. In previous version, there was for
example two `mToday` variable, since both are private. One as
attribute of Sched, and the other one as attribute of SchedV2.