Shere sorted deck collection between _checkDeckTree and deckDueList
On my collection, sorting the list of decks takes 4 seconds. This is
done twice when computing the list of deck (which is required to load
the gui). Ideally, I'd love to cache the sorted list of deck and
delete the cache as soon as the list of deck is changed. I fear it
won't be accepted, since another cache I wanted to add was rejected. I
hope that this compromise would work; the change between my proposal
and upstream anki is really small; it allows to use the sorted list
in the two positions; thus decreasing loading time by 4 seconds (in my
collection, with a lot of deck and subdecks)