Open deck when number of cards is unknwon.
I assume that if the user click on a deck while numbers are not
loaded, we should try to open this deck has if he have card. If it has
no card, it'll come leave the reviewer anyway. Otherwise, it'll save a
lot of time to the user if they does not need the result of the computation.