checking which cards are marked is moved to CardBrowser
Currently, asking the browser to show the whole collection takes
multiple minutes. On a smaller deck, it takes 40 seconds, out of which
15 are spent executing the regexp and looking for marked cards.

Moving it to updateSearchItemQA ensure it is executed only for card we
intend to display and save most of this time.

Since "marked" is computed later, the color of the browser line is
computed at the same time than the text, instead of being
pre-computed. So it makes the following difference:
* before this commit, all lines are directly colored
* after this commit, if you scroll quickly, all lines are white and
  they become colored at the moment when they get their text.

(To be even more precise, the text is filled or not depending on the
columns. "sort field" is always filled while "question" is filled only
after the line appear on screen.)