Expand and Collapse button indicate which deck it acts on
According to the accessibility scanner, it's require to allow the user to know what will get extended.

I assumed that the last deck name component would still be indicative while being short enough to be useful. The long
may be more exhaustive, have no conflict, but less easy to read.