Factorize code that require to confirm modification of schema
This does not touch the cases where the warning is given because of a bug.

This change the behaviour of the "delete note type", "delete field" case. It now always ask for
confirmation. Previously, it only did if the schema was not already modified.

In addFieldDialog, sortByField and repositionFieldDialog, this lead to `dismissContextMenu` being called even if no confirmation was
needed. This causes no problem as a check for null is made in this method.