Testing crowdin
This test check whether all translation information remains after a change in file.

I added French translation: This is a second translation, just to try
and  comment: This is a comment test, to check for moving file

The test is entirely succesful if they are kept in translation.

If the test fails, I'll need to check what occurs if the same change is made in French and English simultaneously (This
will require to change sync file)