NF: Redefine `DeckId` as a value class
This ensure that we can finally benefits from actual type checking as
far for the dids. This work was considerably helped by previous
cleaning which ensured that we mostly used `DeckID` everywhere a long
is supposed to be the id of a deck.

If this commit is accepted, I plan to have similar commits for other
ids.

Ideally, it'd be nice if we could avoid to doing as many type
transformation when saving the deck in a Preference, Bundle or backend.